Correccion de errores y validaciones en Habitaciones y reservas

+++ BackEnd/commons-reservas/src/main/java/com/proyecto/commons/dto/DatosHabitacion.java
@@ -7,6 +7,6 @@ public record DatosHabitacion(
 		Integer numero,
 		String tipo,
 		BigDecimal precio,
-		Integer capacidad
+		Short capacidad
 		
 ) {}
